env-var-unset: cushion the unset command
Currently, if an ENV-variable contains a leading
":", the unset fails with "Incorrect environment
variable name". We check whether an env-var conforms
to a regex. Also, we remove the directories used
in the build, in order to be able to re-launch
the script in case of failures (e.g., one forgot
to launch it in a screen session of the remote)

Signed-off-by: sk <60398690+serkazi@users.noreply.github.com>